moving *test_read_file* to module api updating module api docker
