+++ Packs/ApiModules/Scripts/MLApiModule/TestData/input_json_file_test
@@ -1,0 +1,2 @@
+[{"body": "This is a sample of the body", "subject": "This is a sample of the subject"},
+{"body2": "This is a sample of the body2", "subject2": "This is a sample of the subject2", "subject": "This is a sample of the subject"}]
